pyproject.toml
[project]
name = "empty-django-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-environ>=0.12.0",
    "gunicorn>=23.0.0",
    "whitenoise>=6.11.0",
]
